+++ app/policies/export_policy.rb
@@ -16,6 +16,7 @@ class ExportPolicy
   def csv?
     return true if @user.education_and_skills_funding_agency_team?
     return true if @user.academies_operational_practice_unit_team?
+    return true if @user.business_support_team?
 
     false
   end